requirements.txt
python-telegram-bot==21.1.1
google-genai==0.3.0
PyMuPDF==1.24.1
python-docx==1.1.0
xhtml2pdf==0.2.16
Jinja2==3.1.3
python-dotenv==1.0.1
pydantic==2.7.1
